Made it more convenient to perform a shift/rotate by 1 operation
The value by which a number is usually shifted/rotated is not big, so it doesn't
make much sense to use the same number as the second operand. Set 1 as the new
default and perform shift/rotate by 1 if the same button is clicked more than
once. While the second operand is unchanged, also perform unary operations
(Not, sign swap, load Min/Max) and manual bit changes on the first operand.
By convention, MS, M+ and M- will work with the *first* operand while
MR will replace the *second* operand with the stored value.